cursorproject/requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
pinecone-client==3.0.2
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
langchain==0.1.11
openai==1.12.0
tiktoken==0.6.0
python-multipart==0.0.9
new-package==1.0.0